Address file drop review feedback
