+++ Sources/ContentView.swift
@@ -7,971 +7,6 @@ import ObjectiveC
 import UniformTypeIdentifiers
 import WebKit
 
-private final class FileDropHintBadgeView: NSView {
-    private static let neutralBackgroundColor = NSColor.systemGray.withAlphaComponent(0.20)
-    private let effectView: NSView
-    private let label = NSTextField(labelWithString: "")
-
-    override var isOpaque: Bool { false }
-
-    override init(frame frameRect: NSRect) {
-        if let glassClass = NSClassFromString("NSGlassEffectView") as? NSView.Type {
-            effectView = glassClass.init(frame: .zero)
-        } else {
-            let visualEffect = NSVisualEffectView(frame: .zero)
-            visualEffect.material = .hudWindow
-            visualEffect.blendingMode = .withinWindow
-            visualEffect.state = .active
-            effectView = visualEffect
-        }
-
-        super.init(frame: frameRect)
-        frame.size = CGSize(width: 140, height: 26)
-
-        wantsLayer = true
-        layer?.backgroundColor = NSColor.clear.cgColor
-        alphaValue = 0
-        isHidden = true
-
-        effectView.translatesAutoresizingMaskIntoConstraints = false
-        effectView.wantsLayer = true
-        effectView.layer?.backgroundColor = Self.neutralBackgroundColor.cgColor
-        effectView.layer?.cornerRadius = 13
-        effectView.layer?.masksToBounds = true
-        configureNativeGlassIfNeeded(effectView)
-        addSubview(effectView)
-
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .systemFont(ofSize: 12, weight: .medium)
-        label.textColor = .labelColor
-        label.alignment = .center
-        label.lineBreakMode = .byClipping
-        label.maximumNumberOfLines = 1
-        label.setContentCompressionResistancePriority(.required, for: .horizontal)
-        addSubview(label, positioned: .above, relativeTo: effectView)
-
-        NSLayoutConstraint.activate([
-            effectView.topAnchor.constraint(equalTo: topAnchor),
-            effectView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            effectView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            effectView.trailingAnchor.constraint(equalTo: trailingAnchor),
-
-            label.topAnchor.constraint(equalTo: topAnchor, constant: 5),
-            label.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -5),
-            label.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10),
-            label.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -10),
-        ])
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        nil
-    }
-
-    func show(text: String, centeredIn targetBounds: CGRect, clippedTo bounds: CGRect) {
-        label.stringValue = text
-        let fitting = label.intrinsicContentSize
-        let maxWidth = max(80, min(bounds.width, targetBounds.width) - 16)
-        let width = min(max(140, fitting.width + 20), maxWidth)
-        let height: CGFloat = 26
-        let origin = CGPoint(
-            x: min(max(targetBounds.midX - width / 2, bounds.minX + 8), max(bounds.minX + 8, bounds.maxX - width - 8)),
-            y: min(max(targetBounds.midY - height / 2, bounds.minY + 8), max(bounds.minY + 8, bounds.maxY - height - 8))
-        )
-        frame = CGRect(origin: origin, size: CGSize(width: width, height: height))
-        if isHidden {
-            alphaValue = 0
-            isHidden = false
-            animator().alphaValue = 1
-        } else {
-            alphaValue = 1
-        }
-    }
-
-    func hide() {
-        guard !isHidden else { return }
-        alphaValue = 0
-        isHidden = true
-    }
-
-    private func configureNativeGlassIfNeeded(_ view: NSView) {
-        guard view.className == "NSGlassEffectView" else { return }
-
-        let tintSelector = NSSelectorFromString("setTintColor:")
-        if view.responds(to: tintSelector) {
-            view.perform(tintSelector, with: Self.neutralBackgroundColor)
-        }
-
-        let cornerRadiusSelector = NSSelectorFromString("setCornerRadius:")
-        if view.responds(to: cornerRadiusSelector),
-           let implementation = view.method(for: cornerRadiusSelector) {
-            typealias CornerRadiusSetter = @convention(c) (AnyObject, Selector, CGFloat) -> Void
-            let setter = unsafeBitCast(implementation, to: CornerRadiusSetter.self)
-            setter(view, cornerRadiusSelector, 13)
-        }
-    }
-}
-
-@MainActor
-private protocol FileDropPaneTarget: AnyObject {
-    func fileDropDraggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation
-    func fileDropDraggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation
-    func fileDropDraggingExited(_ sender: (any NSDraggingInfo)?)
-    func fileDropPrepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool
-    func fileDropPerformDragOperation(_ sender: any NSDraggingInfo) -> Bool
-    func fileDropConcludeDragOperation(_ sender: (any NSDraggingInfo)?)
-}
-
-extension PaneDropTargetView: FileDropPaneTarget {
-    func fileDropDraggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation { draggingEntered(sender) }
-    func fileDropDraggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation { draggingUpdated(sender) }
-    func fileDropDraggingExited(_ sender: (any NSDraggingInfo)?) { draggingExited(sender) }
-    func fileDropPrepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool { prepareForDragOperation(sender) }
-    func fileDropPerformDragOperation(_ sender: any NSDraggingInfo) -> Bool { performDragOperation(sender) }
-    func fileDropConcludeDragOperation(_ sender: (any NSDraggingInfo)?) { concludeDragOperation(sender) }
-}
-
-extension BrowserPaneDropTargetView: FileDropPaneTarget {
-    func fileDropDraggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation { draggingEntered(sender) }
-    func fileDropDraggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation { draggingUpdated(sender) }
-    func fileDropDraggingExited(_ sender: (any NSDraggingInfo)?) { draggingExited(sender) }
-    func fileDropPrepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool { prepareForDragOperation(sender) }
-    func fileDropPerformDragOperation(_ sender: any NSDraggingInfo) -> Bool { performDragOperation(sender) }
-    func fileDropConcludeDragOperation(_ sender: (any NSDraggingInfo)?) { concludeDragOperation(sender) }
-}
-
-/// Transparent NSView installed on the window's theme frame (above the NSHostingView) to
-/// handle file/URL drags from Finder. Nested NSHostingController layers (created by bonsplit's
-/// SinglePaneWrapper) prevent AppKit's NSDraggingDestination routing from reaching deeply
-/// embedded terminal views. This overlay sits above the entire content view hierarchy and
-/// intercepts file drags, forwarding drops to the GhosttyNSView under the cursor.
-///
-/// Mouse events are forwarded to the views below via a hide-send-unhide pattern so clicks,
-/// scrolls, and other interactions pass through normally.
-final class FileDropOverlayView: NSView {
-    /// Fallback handler when no terminal is found under the drop point.
-    var onDrop: (([URL]) -> Bool)?
-    private var isForwardingMouseEvent = false
-    private weak var forwardedMouseDragTarget: NSView?
-    private var forwardedMouseDragButton: ForwardedMouseDragButton?
-    /// The WKWebView currently receiving forwarded drag events, so we can
-    /// synthesize draggingExited/draggingEntered as the cursor moves.
-    private weak var activeDragWebView: WKWebView?
-    /// The WKWebView that accepted prepareForDragOperation so conclude can be
-    /// delivered to the same browser target after the drop completes.
-    private weak var preparedDragWebView: WKWebView?
-    /// Pane drop target currently receiving delegated file drag events.
-    private weak var activePaneDropTarget: (any FileDropPaneTarget)?
-    /// Pane drop target that accepted prepareForDragOperation.
-    private weak var preparedPaneDropTarget: (any FileDropPaneTarget)?
-    private let hintBadgeView = FileDropHintBadgeView(frame: .zero)
-    private var lastHitTestLogSignature: String?
-    private var lastDragRouteLogSignatureByPhase: [String: String] = [:]
-
-    override var acceptsFirstResponder: Bool { false }
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        registerForDraggedTypes(Array(PasteboardFileURLReader.fileURLPasteboardTypes))
-        addSubview(hintBadgeView)
-    }
-
-    required init?(coder: NSCoder) { fatalError("init(coder:) not implemented") }
-
-    private enum ForwardedMouseDragButton: Equatable {
-        case left
-        case right
-        case other(Int)
-    }
-
-    private func dragButton(for event: NSEvent) -> ForwardedMouseDragButton? {
-        switch event.type {
-        case .leftMouseDown, .leftMouseUp, .leftMouseDragged:
-            return .left
-        case .rightMouseDown, .rightMouseUp, .rightMouseDragged:
-            return .right
-        case .otherMouseDown, .otherMouseUp, .otherMouseDragged:
-            return .other(Int(event.buttonNumber))
-        default:
-            return nil
-        }
-    }
-
-    private func shouldTrackForwardedMouseDragStart(for eventType: NSEvent.EventType) -> Bool {
-        switch eventType {
-        case .leftMouseDown, .rightMouseDown, .otherMouseDown:
-            return true
-        default:
-            return false
-        }
-    }
-
-    private func shouldTrackForwardedMouseDragEnd(for eventType: NSEvent.EventType) -> Bool {
-        switch eventType {
-        case .leftMouseUp, .rightMouseUp, .otherMouseUp:
-            return true
-        default:
-            return false
-        }
-    }
-
-    // MARK: Hit-testing — participation is routed by DragOverlayRoutingPolicy so
-    // file-drop, bonsplit tab drags, and sidebar tab reorder drags cannot conflict.
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        let pb = NSPasteboard(name: .drag)
-        let eventType = NSApp.currentEvent?.type
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropOverlay(
-            pasteboardTypes: pb.types,
-            eventType: eventType
-        )
-#if DEBUG
-        logHitTestDecision(
-            pasteboardTypes: pb.types,
-            eventType: eventType,
-            shouldCapture: shouldCapture
-        )
-#endif
-        guard shouldCapture else { return nil }
-        if shouldDeferFileDropOverlayToBonsplitTabBar(at: point) {
-            return nil
-        }
-
-        return super.hitTest(point)
-    }
-
-    // MARK: Mouse forwarding — safety net for the rare case where stale drag pasteboard
-    // data causes hitTest to return self when no drag is actually active.
-    // We hit-test contentView directly and dispatch to the target rather than using
-    // window.sendEvent(), which caches the mouse target and causes infinite recursion.
-
-    private func forwardEvent(_ event: NSEvent) {
-        guard !isForwardingMouseEvent else { return }
-        guard let window, let contentView = window.contentView else { return }
-        let eventButton = dragButton(for: event)
-
-        isForwardingMouseEvent = true
-        isHidden = true
-        defer {
-            isHidden = false
-            isForwardingMouseEvent = false
-        }
-
-        let target: NSView?
-        if let eventButton,
-           forwardedMouseDragButton == eventButton,
-           let activeTarget = forwardedMouseDragTarget,
-           activeTarget.window != nil {
-            // Preserve normal AppKit mouse-delivery semantics: once a drag starts,
-            // keep routing dragged/up events to the original mouseDown target.
-            target = activeTarget
-        } else {
-            let point = contentView.convert(event.locationInWindow, from: nil)
-            target = contentView.hitTest(point)
-        }
-
-        guard let target, target !== self else {
-            if shouldTrackForwardedMouseDragEnd(for: event.type),
-               let eventButton,
-               forwardedMouseDragButton == eventButton {
-                forwardedMouseDragTarget = nil
-                forwardedMouseDragButton = nil
-            }
-            return
-        }
-
-        if shouldTrackForwardedMouseDragStart(for: event.type), let eventButton {
-            forwardedMouseDragTarget = target
-            forwardedMouseDragButton = eventButton
-        }
-
-        switch event.type {
-        case .leftMouseDown: target.mouseDown(with: event)
-        case .leftMouseUp: target.mouseUp(with: event)
-        case .leftMouseDragged: target.mouseDragged(with: event)
-        case .rightMouseDown: target.rightMouseDown(with: event)
-        case .rightMouseUp: target.rightMouseUp(with: event)
-        case .rightMouseDragged: target.rightMouseDragged(with: event)
-        case .otherMouseDown: target.otherMouseDown(with: event)
-        case .otherMouseUp: target.otherMouseUp(with: event)
-        case .otherMouseDragged: target.otherMouseDragged(with: event)
-        case .scrollWheel: target.scrollWheel(with: event)
-        default: break
-        }
-
-        if shouldTrackForwardedMouseDragEnd(for: event.type),
-           let eventButton,
-           forwardedMouseDragButton == eventButton {
-            forwardedMouseDragTarget = nil
-            forwardedMouseDragButton = nil
-        }
-    }
-
-    override func mouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func mouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func mouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func scrollWheel(with event: NSEvent) { forwardEvent(event) }
-
-    // MARK: NSDraggingDestination – accept file drops over terminal and browser views.
-    //
-    // AppKit sends draggingEntered once when the drag enters this overlay, then
-    // draggingUpdated as the cursor moves within it. We track which WKWebView (if
-    // any) is under the cursor and synthesize enter/exit calls so the browser's
-    // HTML5 drag events (dragenter, dragleave, drop) fire correctly.
-
-    override func draggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        return updateDragTarget(sender, phase: "entered")
-    }
-
-    override func draggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        return updateDragTarget(sender, phase: "updated")
-    }
-
-    override func draggingExited(_ sender: (any NSDraggingInfo)?) {
-        hintBadgeView.hide()
-        preparedDragWebView = nil
-        preparedPaneDropTarget = nil
-        if let prev = activeDragWebView {
-            prev.draggingExited(sender)
-            activeDragWebView = nil
-        }
-        if let prev = activePaneDropTarget {
-            prev.fileDropDraggingExited(sender)
-            activePaneDropTarget = nil
-        }
-    }
-
-    override func prepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        if shouldRouteFileDropToTextDestination(sender) {
-            preparedDragWebView = nil
-            if let paneDropTarget = activePaneDropTarget ?? paneDropTargetForTextDrop(at: sender.draggingLocation) {
-                let accepted = paneDropTarget.fileDropPrepareForDragOperation(sender)
-                preparedPaneDropTarget = accepted ? paneDropTarget : nil
-                return accepted
-            }
-            preparedPaneDropTarget = nil
-            if let webView = webViewUnderPoint(sender.draggingLocation) {
-                let accepted = webView.prepareForDragOperation(sender)
-                preparedDragWebView = accepted ? webView : nil
-                return accepted
-            }
-            return textDropDestinationKindUnderPoint(sender.draggingLocation) != nil
-        }
-        let paneDropTarget = shouldCapture
-            ? (activePaneDropTarget ?? paneDropTargetUnderPoint(sender.draggingLocation))
-            : nil
-        let terminal = paneDropTarget == nil ? terminalUnderPoint(sender.draggingLocation) : nil
-        let webView = paneDropTarget == nil && terminal == nil ? webViewUnderPoint(sender.draggingLocation) : nil
-        let hasPaneTarget = paneDropTarget != nil || terminal != nil || webView != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: "prepare",
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasPaneTarget: hasPaneTarget
-        )
-#endif
-        guard shouldCapture else {
-            preparedDragWebView = nil
-            preparedPaneDropTarget = nil
-            activePaneDropTarget = nil
-            return false
-        }
-        preparedDragWebView = nil
-        if let paneDropTarget {
-            let accepted = paneDropTarget.fileDropPrepareForDragOperation(sender)
-            preparedPaneDropTarget = accepted ? paneDropTarget : nil
-            return accepted
-        }
-        preparedPaneDropTarget = nil
-        if let webView {
-            let accepted = webView.prepareForDragOperation(sender)
-            preparedDragWebView = accepted ? webView : nil
-            return accepted
-        }
-        return hasPaneTarget
-    }
-
-    override func performDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        if shouldRouteFileDropToTextDestination(sender) {
-            hintBadgeView.hide()
-            preparedDragWebView = nil
-            if let paneDropTarget = preparedPaneDropTarget ?? activePaneDropTarget ?? paneDropTargetForTextDrop(at: sender.draggingLocation) {
-                let handled = paneDropTarget.fileDropPerformDragOperation(sender)
-                if !handled {
-                    preparedPaneDropTarget = nil
-                    activePaneDropTarget = nil
-                }
-                return handled
-            }
-            if let webView = preparedDragWebView ?? webViewUnderPoint(sender.draggingLocation) {
-                let handled = webView.performDragOperation(sender)
-                if !handled {
-                    preparedDragWebView = nil
-                    activeDragWebView = nil
-                }
-                return handled
-            }
-            return performFileDropAsText(sender)
-        }
-        let paneDropTarget = shouldCapture
-            ? (preparedPaneDropTarget ?? activePaneDropTarget ?? paneDropTargetUnderPoint(sender.draggingLocation))
-            : nil
-        let terminal = paneDropTarget == nil ? terminalUnderPoint(sender.draggingLocation) : nil
-        let webView = paneDropTarget == nil && terminal == nil
-            ? (preparedDragWebView ?? activeDragWebView ?? webViewUnderPoint(sender.draggingLocation))
-            : nil
-        let hasPaneTarget = paneDropTarget != nil || terminal != nil || webView != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: "perform",
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasPaneTarget: hasPaneTarget
-        )
-#endif
-        guard shouldCapture else {
-            preparedDragWebView = nil
-            activeDragWebView = nil
-            preparedPaneDropTarget = nil
-            activePaneDropTarget = nil
-            return false
-        }
-        preparedDragWebView = nil
-        if let paneDropTarget {
-            let handled = paneDropTarget.fileDropPerformDragOperation(sender)
-            if !handled {
-                preparedPaneDropTarget = nil
-                activePaneDropTarget = nil
-            }
-            return handled
-        }
-        preparedPaneDropTarget = nil
-        if let webView {
-            let handled = webView.performDragOperation(sender)
-            if !handled {
-                preparedDragWebView = nil
-                activeDragWebView = nil
-            }
-            return handled
-        }
-        activeDragWebView = nil
-        guard let terminal else { return false }
-        return terminal.performDragOperation(sender)
-    }
-
-    override func concludeDragOperation(_ sender: (any NSDraggingInfo)?) {
-        defer {
-            hintBadgeView.hide()
-            preparedDragWebView = nil
-            activeDragWebView = nil
-            preparedPaneDropTarget = nil
-            activePaneDropTarget = nil
-        }
-        guard let sender else { return }
-        if shouldRouteFileDropToTextDestination(sender) {
-            if let webView = preparedDragWebView ?? activeDragWebView {
-                webView.concludeDragOperation(sender)
-            }
-            return
-        }
-        guard DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: sender.draggingPasteboard.types,
-            hasLocalDraggingSource: sender.draggingSource != nil
-        ) else {
-            return
-        }
-        if let paneDropTarget = preparedPaneDropTarget ?? activePaneDropTarget {
-            paneDropTarget.fileDropConcludeDragOperation(sender)
-            return
-        }
-        if let webView = preparedDragWebView ?? activeDragWebView {
-            webView.concludeDragOperation(sender)
-        }
-    }
-
-    private func updateDragTarget(_ sender: any NSDraggingInfo, phase: String) -> NSDragOperation {
-        let loc = sender.draggingLocation
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        updateHintBadge(sender: sender, pasteboardTypes: types)
-
-        if shouldRouteFileDropToTextDestination(sender) {
-            let paneDropTarget = paneDropTargetForTextDrop(at: loc)
-            if let prev = activePaneDropTarget {
-                if fileDropPaneTargetsAreIdentical(prev, paneDropTarget) {
-                    return prev.fileDropDraggingUpdated(sender)
-                }
-                prev.fileDropDraggingExited(sender)
-                activePaneDropTarget = nil
-            }
-            if let paneDropTarget {
-                if let prev = activeDragWebView {
-                    prev.draggingExited(sender)
-                    activeDragWebView = nil
-                }
-                activePaneDropTarget = paneDropTarget
-                return paneDropTarget.fileDropDraggingEntered(sender)
-            }
-            if let webView = webViewUnderPoint(loc) {
-                if activeDragWebView !== webView {
-                    if let prev = activeDragWebView {
-                        prev.draggingExited(sender)
-                    }
-                    activeDragWebView = webView
-                    return webView.draggingEntered(sender)
-                }
-                return webView.draggingUpdated(sender)
-            }
-            if let prev = activeDragWebView {
-                prev.draggingExited(sender)
-                activeDragWebView = nil
-            }
-            return textDropDestinationKindUnderPoint(loc) == nil
-                ? []
-                : DragOverlayRoutingPolicy.textDropOperation(pasteboardTypes: types)
-        }
-
-        let paneDropTarget = shouldCapture ? paneDropTargetUnderPoint(loc) : nil
-        let webView = shouldCapture && paneDropTarget == nil ? webViewUnderPoint(loc) : nil
-
-        if let prev = activeDragWebView {
-            if prev !== webView {
-                prev.draggingExited(sender)
-                activeDragWebView = nil
-            }
-        }
-        if let prev = activePaneDropTarget,
-           !fileDropPaneTargetsAreIdentical(prev, paneDropTarget) {
-            prev.fileDropDraggingExited(sender)
-            activePaneDropTarget = nil
-        }
-
-        if let paneDropTarget {
-            if !fileDropPaneTargetsAreIdentical(activePaneDropTarget, paneDropTarget) {
-                activePaneDropTarget = paneDropTarget
-                return paneDropTarget.fileDropDraggingEntered(sender)
-            }
-            return paneDropTarget.fileDropDraggingUpdated(sender)
-        }
-
-        if let webView {
-            if activeDragWebView !== webView {
-                activeDragWebView = webView
-                return webView.draggingEntered(sender)
-            }
-            return webView.draggingUpdated(sender)
-        }
-
-        let hasPaneTarget = terminalUnderPoint(loc) != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: phase,
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasPaneTarget: hasPaneTarget
-        )
-#endif
-        guard shouldCapture, hasPaneTarget else { return [] }
-        return .copy
-    }
-
-    private func fileDropPaneTargetsAreIdentical(
-        _ lhs: (any FileDropPaneTarget)?,
-        _ rhs: (any FileDropPaneTarget)?
-    ) -> Bool {
-        guard let lhs, let rhs else { return lhs == nil && rhs == nil }
-        return (lhs as AnyObject) === (rhs as AnyObject)
-    }
-
-    private func debugPasteboardTypes(_ types: [NSPasteboard.PasteboardType]?) -> String {
-        guard let types, !types.isEmpty else { return "-" }
-        return types.map(\.rawValue).joined(separator: ",")
-    }
-
-    private func shouldRouteFileDropToTextDestination(_ sender: any NSDraggingInfo) -> Bool {
-        let canDropAsText = textDropDestinationKindUnderPoint(sender.draggingLocation) != nil
-        return DragOverlayRoutingPolicy.shouldRouteFileDropToTextDestination(
-            pasteboardTypes: sender.draggingPasteboard.types,
-            modifierFlags: DragOverlayRoutingPolicy.currentModifierFlags,
-            canDropAsText: canDropAsText
-        )
-    }
-
-    private func updateHintBadge(
-        sender: any NSDraggingInfo,
-        pasteboardTypes: [NSPasteboard.PasteboardType]?
-    ) {
-        let kind = textDropDestinationKindUnderPoint(sender.draggingLocation)
-        guard let alternateBehavior = DragOverlayRoutingPolicy.alternateFileDropBehaviorForShiftHint(
-            pasteboardTypes: pasteboardTypes,
-            modifierFlags: DragOverlayRoutingPolicy.currentModifierFlags,
-            canDropAsText: kind != nil
-        ), let kind,
-           let hintText = kind.hintText(for: alternateBehavior),
-           let targetBounds = hintBadgeTargetBoundsUnderPoint(sender.draggingLocation) else {
-            hintBadgeView.hide()
-            return
-        }
-        hintBadgeView.show(text: hintText, centeredIn: targetBounds, clippedTo: bounds)
-    }
-
-    private func textDropDestinationKindUnderPoint(_ windowPoint: NSPoint) -> FileDropTextDestinationKind? {
-        if editableTextViewUnderPoint(windowPoint) != nil {
-            return .editor
-        }
-        if webViewUnderPoint(windowPoint) != nil {
-            return .editor
-        }
-        if terminalUnderPoint(windowPoint) != nil {
-            return .terminal
-        }
-        return nil
-    }
-
-    private func hintBadgeTargetBoundsUnderPoint(_ windowPoint: NSPoint) -> CGRect? {
-        if let paneDropTarget = paneDropTargetUnderPoint(windowPoint),
-           let targetView = paneDropTarget as? NSView {
-            return convert(targetView.bounds, from: targetView)
-        }
-        if let terminal = terminalUnderPoint(windowPoint) {
-            return convert(terminal.bounds, from: terminal)
-        }
-        if let webView = webViewUnderPoint(windowPoint) {
-            return convert(webView.bounds, from: webView)
-        }
-        if let textView = editableTextViewUnderPoint(windowPoint) {
-            return convert(textView.visibleRect, from: textView)
-        }
-        return nil
-    }
-
-    private func performFileDropAsText(_ sender: any NSDraggingInfo) -> Bool {
-        let urls = DragOverlayRoutingPolicy.fileURLs(from: sender.draggingPasteboard)
-        guard !urls.isEmpty else { return false }
-        let text = TerminalImageTransferPlanner.insertedText(forFileURLs: urls)
-        guard !text.isEmpty else { return false }
-
-        let windowPoint = sender.draggingLocation
-        if let textView = editableTextViewUnderPoint(windowPoint) {
-            return insert(text, into: textView)
-        }
-        if let webView = webViewUnderPoint(windowPoint) {
-            return insert(text, into: webView, at: windowPoint)
-        }
-        if let terminal = terminalUnderPoint(windowPoint) {
-            return insert(urls, into: terminal)
-        }
-        return false
-    }
-
-    private func viewUnderPoint(_ windowPoint: NSPoint) -> NSView? {
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-        let point = contentView.convert(windowPoint, from: nil)
-        return contentView.hitTest(point)
-    }
-
-    private func editableTextViewUnderPoint(_ windowPoint: NSPoint) -> NSTextView? {
-        var current = viewUnderPoint(windowPoint)
-        while let view = current {
-            if let textView = view as? NSTextView, textView.isEditable {
-                return textView
-            }
-            if let textField = view as? NSTextField,
-               textField.isEditable,
-               let editor = textField.currentEditor() as? NSTextView {
-                return editor
-            }
-            current = view.superview
-        }
-
-        if let fieldEditor = window?.firstResponder as? NSTextView,
-           fieldEditor.isFieldEditor,
-           fieldEditor.isEditable {
-            return fieldEditor
-        }
-        return nil
-    }
-
-    private func insert(_ text: String, into textView: NSTextView) -> Bool {
-        guard textView.isEditable else { return false }
-        textView.window?.makeFirstResponder(textView)
-        textView.insertText(text, replacementRange: textView.selectedRange())
-        return true
-    }
-
-    private func insert(_ text: String, into webView: WKWebView, at windowPoint: NSPoint) -> Bool {
-        FileDropTextInsertion.insert(text, into: webView, at: windowPoint)
-    }
-
-    private func insert(_ urls: [URL], into terminal: GhosttyNSView) -> Bool {
-        let handled = terminal.handleDroppedFileURLsAsText(urls)
-        guard handled,
-              let workspaceId = terminal.tabId,
-              let panelId = terminal.terminalSurface?.id,
-              let workspace = AppDelegate.shared?.workspaceFor(tabId: workspaceId) else {
-            return handled
-        }
-        FileDropTextDropController.focusPanelAfterSuccessfulTextDrop(
-            workspace: workspace,
-            panelId: panelId,
-            focusIntent: .terminal(.surface),
-            window: terminal.window
-        )
-        return true
-    }
-
-    /// Hit-tests the window to find a WKWebView (browser panel) under the cursor.
-    func webViewUnderPoint(_ windowPoint: NSPoint) -> WKWebView? {
-        if let window,
-           let portalWebView = BrowserWindowPortalRegistry.webViewAtWindowPoint(windowPoint, in: window) {
-            return portalWebView
-        }
-
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-        let point = contentView.convert(windowPoint, from: nil)
-        let hitView = contentView.hitTest(point)
-
-        var current: NSView? = hitView
-        while let view = current {
-            if let webView = view as? WKWebView { return webView }
-            current = view.superview
-        }
-        return nil
-    }
-
-    private func debugTopHitViewForCurrentEvent() -> String {
-        guard let window,
-              let currentEvent = NSApp.currentEvent,
-              let contentView = window.contentView,
-              let themeFrame = contentView.superview else { return "-" }
-
-        let pointInTheme = themeFrame.convert(currentEvent.locationInWindow, from: nil)
-        // Don't toggle isHidden here — it triggers setNeedsDisplay which can
-        // exceed AppKit's display-pass limit during cursor-update display cycles.
-        guard let hit = themeFrame.hitTest(pointInTheme) else { return "nil" }
-        var chain: [String] = []
-        var current: NSView? = hit
-        var depth = 0
-        while let view = current, depth < 6 {
-            chain.append(debugHitViewDescriptor(view))
-            current = view.superview
-            depth += 1
-        }
-        return chain.joined(separator: "->")
-    }
-
-    private func debugHitViewDescriptor(_ view: NSView) -> String {
-        let className = String(describing: type(of: view))
-        let ptr = String(describing: Unmanaged.passUnretained(view).toOpaque())
-        let dragTypes = debugRegisteredDragTypes(view)
-        return "\(className)@\(ptr){dragTypes=\(dragTypes)}"
-    }
-
-    private func debugRegisteredDragTypes(_ view: NSView) -> String {
-        let types = view.registeredDraggedTypes
-        guard !types.isEmpty else { return "-" }
-
-        let interestingTypes = types.filter { type in
-            let raw = type.rawValue
-            return PasteboardFileURLReader.fileURLPasteboardTypes.contains(type)
-                || raw == DragOverlayRoutingPolicy.bonsplitTabTransferType.rawValue
-                || raw == DragOverlayRoutingPolicy.sidebarTabReorderType.rawValue
-                || raw.contains("public.text")
-                || raw.contains("public.url")
-                || raw.contains("public.data")
-        }
-        let selected = interestingTypes.isEmpty ? Array(types.prefix(3)) : interestingTypes
-        let rendered = selected.map(\.rawValue).joined(separator: ",")
-        if selected.count < types.count {
-            return "\(rendered),+\(types.count - selected.count)"
-        }
-        return rendered
-    }
-
-    private func hasRelevantDragTypes(_ types: [NSPasteboard.PasteboardType]?) -> Bool {
-        guard let types else { return false }
-        return DragOverlayRoutingPolicy.hasFileDropPayload(types)
-            || types.contains(DragOverlayRoutingPolicy.bonsplitTabTransferType)
-            || types.contains(DragOverlayRoutingPolicy.sidebarTabReorderType)
-    }
-
-    private func debugEventName(_ eventType: NSEvent.EventType?) -> String {
-        guard let eventType else { return "none" }
-        switch eventType {
-        case .cursorUpdate: return "cursorUpdate"
-        case .appKitDefined: return "appKitDefined"
-        case .systemDefined: return "systemDefined"
-        case .applicationDefined: return "applicationDefined"
-        case .periodic: return "periodic"
-        case .mouseMoved: return "mouseMoved"
-        case .mouseEntered: return "mouseEntered"
-        case .mouseExited: return "mouseExited"
-        case .flagsChanged: return "flagsChanged"
-        case .leftMouseDown: return "leftMouseDown"
-        case .leftMouseUp: return "leftMouseUp"
-        case .leftMouseDragged: return "leftMouseDragged"
-        case .rightMouseDown: return "rightMouseDown"
-        case .rightMouseUp: return "rightMouseUp"
-        case .rightMouseDragged: return "rightMouseDragged"
-        case .otherMouseDown: return "otherMouseDown"
-        case .otherMouseUp: return "otherMouseUp"
-        case .otherMouseDragged: return "otherMouseDragged"
-        case .scrollWheel: return "scrollWheel"
-        default: return "other(\(eventType.rawValue))"
-        }
-    }
-
-#if DEBUG
-    private func logHitTestDecision(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        eventType: NSEvent.EventType?,
-        shouldCapture: Bool
-    ) {
-        let isDragEvent = eventType == .leftMouseDragged
-            || eventType == .rightMouseDragged
-            || eventType == .otherMouseDragged
-        guard shouldCapture || isDragEvent || hasRelevantDragTypes(pasteboardTypes) else { return }
-
-        let signature = "\(shouldCapture ? 1 : 0)|\(debugEventName(eventType))|\(debugPasteboardTypes(pasteboardTypes))"
-        guard lastHitTestLogSignature != signature else { return }
-        lastHitTestLogSignature = signature
-        cmuxDebugLog(
-            "overlay.fileDrop.hitTest capture=\(shouldCapture ? 1 : 0) " +
-            "event=\(debugEventName(eventType)) " +
-            "topHit=\(debugTopHitViewForCurrentEvent()) " +
-            "types=\(debugPasteboardTypes(pasteboardTypes))"
-        )
-    }
-
-    private func logDragRouteDecision(
-        phase: String,
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        shouldCapture: Bool,
-        hasLocalDraggingSource: Bool,
-        hasPaneTarget: Bool
-    ) {
-        guard shouldCapture || hasRelevantDragTypes(pasteboardTypes) else { return }
-        let signature = [
-            shouldCapture ? "1" : "0",
-            hasLocalDraggingSource ? "1" : "0",
-            hasPaneTarget ? "1" : "0",
-            debugPasteboardTypes(pasteboardTypes)
-        ].joined(separator: "|")
-        guard lastDragRouteLogSignatureByPhase[phase] != signature else { return }
-        lastDragRouteLogSignatureByPhase[phase] = signature
-        cmuxDebugLog(
-            "overlay.fileDrop.\(phase) capture=\(shouldCapture ? 1 : 0) " +
-            "localSource=\(hasLocalDraggingSource ? 1 : 0) " +
-            "hasPane=\(hasPaneTarget ? 1 : 0) " +
-            "types=\(debugPasteboardTypes(pasteboardTypes))"
-        )
-    }
-#endif
-    /// Hit-tests the window to find the GhosttyNSView under the cursor.
-    func terminalUnderPoint(_ windowPoint: NSPoint) -> GhosttyNSView? {
-        if let window,
-           let portalTerminal = TerminalWindowPortalRegistry.terminalViewAtWindowPoint(windowPoint, in: window) {
-            return portalTerminal
-        }
-
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-        let point = contentView.convert(windowPoint, from: nil)
-        let hitView = contentView.hitTest(point)
-
-        var current: NSView? = hitView
-        while let view = current {
-            if let terminal = view as? GhosttyNSView { return terminal }
-            current = view.superview
-        }
-        return nil
-    }
-
-    private func shouldDeferFileDropOverlayToBonsplitTabBar(at point: NSPoint) -> Bool {
-        guard let window else { return false }
-        let windowPoint = convert(point, to: nil)
-        return BonsplitTabBarHitRegionRegistry.containsWindowPoint(windowPoint, in: window)
-    }
-
-    private func paneDropTargetUnderPoint(_ windowPoint: NSPoint) -> (any FileDropPaneTarget)? {
-        if let paneTarget = inlinePaneDropTargetUnderPoint(windowPoint) {
-            return paneTarget
-        }
-        guard let window else { return nil }
-        if let terminalPaneTarget = TerminalWindowPortalRegistry.terminalPaneDropTargetAtWindowPoint(windowPoint, in: window) {
-            return terminalPaneTarget
-        }
-        return BrowserWindowPortalRegistry.browserPaneDropTargetAtWindowPoint(windowPoint, in: window)
-    }
-
-    private func paneDropTargetForTextDrop(at windowPoint: NSPoint) -> (any FileDropPaneTarget)? {
-        if let textView = editableTextViewUnderPoint(windowPoint),
-           !(textView is SavingTextView) {
-            return nil
-        }
-        return paneDropTargetUnderPoint(windowPoint)
-    }
-
-    private func inlinePaneDropTargetUnderPoint(_ windowPoint: NSPoint) -> PaneDropTargetView? {
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-
-        let point = contentView.convert(windowPoint, from: nil)
-        return paneDropTarget(in: contentView, at: point)
-    }
-
-    private func paneDropTarget(in view: NSView, at point: NSPoint) -> PaneDropTargetView? {
-        for subview in view.subviews.reversed() {
-            guard !subview.isHidden, subview.alphaValue > 0 else { continue }
-            let pointInSubview = subview.convert(point, from: view)
-            guard subview.bounds.contains(pointInSubview) else { continue }
-            if let paneTarget = subview as? PaneDropTargetView {
-                return paneTarget
-            }
-            if let nestedTarget = paneDropTarget(in: subview, at: pointInSubview) {
-                return nestedTarget
-            }
-        }
-        return view as? PaneDropTargetView
-    }
-}
-
 var fileDropOverlayKey: UInt8 = 0
 private var commandPaletteWindowOverlayKey: UInt8 = 0
 private var tmuxWorkspacePaneWindowOverlayKey: UInt8 = 0
